Refined settings dialog and integrated settings dependency into caching.
